Move TableOfContents node into sub-namespace
