+++ src/Extension/TableOfContents/Normalizer/FlatNormalizerStrategy.php
@@ -12,7 +12,7 @@
 namespace League\CommonMark\Extension\TableOfContents\Normalizer;
 
 use League\CommonMark\Extension\CommonMark\Node\Block\ListItem;
-use League\CommonMark\Extension\TableOfContents\TableOfContents;
+use League\CommonMark\Extension\TableOfContents\Node\TableOfContents;
 
 final class FlatNormalizerStrategy implements NormalizerStrategyInterface
 {
